requirements.txt
langchain-community==0.2.7
langgraph==0.1.5
openai==1.35.13
pip-chill==1.0.3
python-dotenv==1.0.1
tavily-python==0.3.3
